# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IoT WebSocket Server API",
    "description": "REST API для управления IoT устройствами (ESP32) с системой автоматического полива. Поддерживает управление конфигурацией, состоянием и всеми компонентами устройств (irrigators, outputs, sensors, timers, pcfOutputs).",
    "version": "2.0.0",
    "contact": {
      "name": "IfNoise",
      "url": "https://github.com/IfNoise/websocket"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3600/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/devices/{deviceId}/call": {
      "post": {
        "tags": ["Device"],
        "summary": "Универсальный RPC вызов",
        "description": "Выполняет произвольный RPC метод на устройстве",
        "operationId": "callDeviceMethod",
        "parameters": [
          {
            "$ref": "#/components/parameters/DeviceId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["method"],
                "properties": {
                  "method": {
                    "type": "string",
                    "description": "Название RPC метода",
                    "example": "Get.Sensors"
                  },
                  "params": {
                    "type": "object",
                    "description": "Параметры метода",
                    "default": {}
                  }
                }
              },
              "examples": {
                "getSensors": {
                  "summary": "Получить данные сенсоров",
                  "value": {
                    "method": "Get.Sensors",
                    "params": {}
                  }
                },
                "setOutput": {
                  "summary": "Установить состояние выхода",
                  "value": {
                    "method": "Set.Output",
                    "params": {
                      "key": "out1",
                      "state": true
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "description": "Результат выполнения метода"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "DeviceId": {
        "name": "deviceId",
        "in": "path",
        "required": true,
        "description": "Уникальный идентификатор устройства",
        "schema": {
          "type": "string",
          "example": "esp32_A8A154"
        }
      },
...
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "Ресурс не найден",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Device not found"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Ошибка валидации",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Внутренняя ошибка сервера",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}